refactor : 문자열을 정수 배열로 변환하는 기능 객체 분리
