+++ src/main/java/StringParser.java
@@ -1,0 +1,4 @@
+public interface StringParser<T> {
+
+    T[] parse(String input);
+}
